# docs/openapi.yml
openapi: 3.0.3
info:
  title: SmartTeach API
  version: 0.2.0
  description: |-
    Актуальная спецификация REST API для MVP проекта SmartTeach.
    Спецификация обновлена после пересмотра предметной области и синхронизирована
    с новой структурой базы данных, зафиксированной в `docs/07-database-structure.md`.

    Основные принципы текущего API:
    - центральная сущность системы — группа;
    - `class` и `course` в MVP не выделяются;
    - у группы есть отдельные настройки функций;
    - вступление в группу зависит от `accessMode`;
    - используются локальная авторизация и refresh-сессии;
    - поддерживаются групповые и личные чаты `1 на 1`;
    - расписание строится из уроков, дедлайнов заданий и кастомных событий.
servers:
  - url: http://localhost:3000/api/v1
    description: Локальный REST API
...
paths:
...
  /groups:
    get:
      description: Возвращает публичные группы без входа в аккаунт и дополняет ответ данными участия, если передан access token.
      parameters:
        - in: query
          name: search
          required: false
          schema:
            example: patterns
            type: string
        - in: query
          name: code
          required: false
          schema:
            example: WEBSPR1
            type: string
        - in: query
          name: accessMode
          required: false
          schema:
            enum:
              - OPEN
              - BY_REQUEST
              - CLOSED
            type: string
        - in: query
          name: status
          required: false
          schema:
            enum:
              - ACTIVE
              - ARCHIVED
              - DELETED
            type: string
        - in: query
          name: joinedOnly
          required: false
          schema:
            example: true
            type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Получить список групп
      tags:
        - Groups
    post:
      description: Создает группу, настройки группы и запись владельца в составе участников.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Создать группу
      tags:
        - Groups
...
  /groups/{groupId}/materials:
    get:
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaterialSection'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Получить структуру материалов группы
      tags:
        - Materials
  /groups/{groupId}/materials/sections:
    post:
      description: Доступно владельцу и администраторам группы, если lessons_enabled включен.
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaterialSectionRequest'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialSection'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Создать раздел материалов
      tags:
        - Materials
  /groups/{groupId}/materials/sections/{sectionId}/subsections:
    post:
      description: Доступно владельцу и администраторам группы, если lessons_enabled включен.
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: sectionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMaterialSubsectionRequest'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialSubsectionDetails'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Создать подраздел материалов
      tags:
        - Materials
  /groups/{groupId}/materials/subsections/{subsectionId}:
    get:
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: subsectionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialSubsectionDetails'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Получить подраздел материалов с уроками
      tags:
        - Materials
...
components:
  securitySchemes:
    bearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 150
          example: API Patterns Lab
        description:
          type: string
          example: Группа для разбора компонентного мышления и интерфейсных паттернов.
        accessMode:
          type: string
          enum:
            - OPEN
            - BY_REQUEST
            - CLOSED
          example: BY_REQUEST
        settings:
          $ref: '#/components/schemas/GroupSettings'
      required:
        - name
        - accessMode
        - settings
...
    CreateMaterialSectionRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 200
          example: HTML и структура страницы
        sortOrder:
          type: number
          minimum: 1
          example: 1
      required:
        - title
    CreateMaterialSubsectionRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 200
          example: Семантическая верстка
        sortOrder:
          type: number
          minimum: 1
          example: 1
      required:
        - title
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          type: string
          example: Validation failed
        errors:
          example:
            - 'email: Invalid email address'
          type: array
          items:
            type: string
      required:
        - statusCode
        - message
...
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 66666666-6666-4666-8666-666666666666
        code:
          type: string
          example: WEBSPR1
        name:
          type: string
          minLength: 2
          maxLength: 150
          example: Web Basics Spring 2026
        description:
          type: object
          nullable: true
          example: Открытая группа для отработки HTML, CSS и базового JavaScript.
        ownerId:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        owner:
          $ref: '#/components/schemas/PublicUser'
        accessMode:
          type: string
          enum:
            - OPEN
            - BY_REQUEST
            - CLOSED
          example: OPEN
        status:
          type: string
          enum:
            - ACTIVE
            - ARCHIVED
            - DELETED
          example: ACTIVE
        settings:
          $ref: '#/components/schemas/GroupSettings'
        membersCount:
          type: number
          minimum: 0
          example: 3
        viewerMembershipRole:
          type: string
          enum:
            - OWNER
            - ADMIN
            - USER
          nullable: true
          description: Роль текущего пользователя в группе, если он уже состоит в ней.
          example: USER
        viewerJoinRequestStatus:
          type: string
          enum:
            - PENDING
            - APPROVED
            - REJECTED
          nullable: true
          description: Последний известный статус заявки текущего пользователя в эту группу.
          example: PENDING
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        archivedAt:
          type: object
          format: date-time
          nullable: true
        deletedAt:
          type: object
          format: date-time
          nullable: true
      required:
        - id
        - code
        - name
        - ownerId
        - owner
        - accessMode
        - status
        - settings
        - membersCount
        - createdAt
        - updatedAt
...
    GroupSettings:
      type: object
      properties:
        chatEnabled:
          type: boolean
          example: true
        lessonsEnabled:
          type: boolean
          example: true
        assignmentsEnabled:
          type: boolean
          example: true
        scheduleEnabled:
          type: boolean
          example: true
        usefulLinksEnabled:
          type: boolean
          example: true
      required:
        - chatEnabled
        - lessonsEnabled
        - assignmentsEnabled
        - scheduleEnabled
        - usefulLinksEnabled
...
    MaterialLesson:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 77777777-7777-4777-8777-777777777777
        groupId:
          type: string
          format: uuid
          example: 66666666-6666-4666-8666-666666666666
        materialSubsectionId:
          type: object
          format: uuid
          nullable: true
          example: 99999999-9999-4999-8999-999999999999
        title:
          type: string
          minLength: 2
          maxLength: 200
          example: Введение в структуру страницы
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - ARCHIVED
          example: PUBLISHED
        sortOrder:
          type: number
          minimum: 1
          example: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - title
        - status
        - sortOrder
        - createdAt
        - updatedAt
    MaterialSection:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 88888888-8888-4888-8888-888888888888
        groupId:
          type: string
          format: uuid
          example: 66666666-6666-4666-8666-666666666666
        title:
          type: string
          minLength: 2
          maxLength: 200
          example: HTML и структура страницы
        sortOrder:
          type: number
          minimum: 1
          example: 1
        createdByUserId:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        subsections:
          type: array
          items:
            $ref: '#/components/schemas/MaterialSubsection'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - title
        - sortOrder
        - createdByUserId
        - subsections
        - createdAt
        - updatedAt
    MaterialSectionSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 88888888-8888-4888-8888-888888888888
        title:
          type: string
          example: HTML и структура страницы
        sortOrder:
          type: number
          minimum: 1
          example: 1
      required:
        - id
        - title
        - sortOrder
    MaterialSubsection:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 99999999-9999-4999-8999-999999999999
        groupId:
          type: string
          format: uuid
          example: 66666666-6666-4666-8666-666666666666
        sectionId:
          type: string
          format: uuid
          example: 88888888-8888-4888-8888-888888888888
        title:
          type: string
          minLength: 2
          maxLength: 200
          example: Семантическая верстка
        sortOrder:
          type: number
          minimum: 1
          example: 1
        lessonsCount:
          type: number
          minimum: 0
          example: 3
        createdByUserId:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - sectionId
        - title
        - sortOrder
        - lessonsCount
        - createdByUserId
        - createdAt
        - updatedAt
    MaterialSubsectionDetails:
      type: object
      properties:
        section:
          $ref: '#/components/schemas/MaterialSectionSummary'
        subsection:
          $ref: '#/components/schemas/MaterialSubsection'
        lessons:
          type: array
          items:
            $ref: '#/components/schemas/MaterialLesson'
      required:
        - section
        - subsection
        - lessons
...
    PublicUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        displayName:
          type: string
          example: Alex Teacher
        bio:
          type: object
          nullable: true
          example: Преподает веб-разработку и ведет открытые учебные группы.
        avatarFileId:
          type: object
          format: uuid
          nullable: true
          example: aaaaaaa1-aaaa-4aaa-8aaa-aaaaaaaaaaa1
        avatarUrl:
          type: object
          format: uri
          nullable: true
          example: http://localhost:9000/smarteach-files/avatars/2026/04/12/aBcD1234xYz9.png?X-Amz-Algorithm=AWS4-HMAC-SHA256
      required:
        - id
        - displayName
...